Move submitted proposal listing to committee module.
Also simplify inheritance, eliminate one level.
