+++ opengever/meeting/configure.zcml
@@ -60,7 +60,6 @@
   <adapter factory=".tabs.meetinglisting.MeetingTableSource" />
   <adapter factory=".tabs.memberlisting.MemberTableSource" />
   <adapter factory=".tabs.membershiplisting.MembershipTableSource" />
-  <adapter factory=".tabs.submittedproposallisting.SubmittedProposalTableSource" />
 
   <adapter factory=".menu.CommitteePostFactoryMenu" />
 </configure>
